filterList logic move to the viewModel, code refactored and removed un used import
